220216 COMPETE

+++ src/boards/boards.service.ts
@@ -5,6 +5,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { BoardRepository } from './board.repository';
 import { Board } from './board.entity';
 import { BoardStatus } from './board.model';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class BoardsService {
@@ -20,13 +21,18 @@ export class BoardsService {
   }
 
   // 게시물 만들기
-  async createBoard(createBoardDto: CreateBoardDto): Promise<Board> {
-    return this.boardRepository.createBoard(createBoardDto);
+  async createBoard(
+    createBoardDto: CreateBoardDto,
+    user: User,
+  ): Promise<Board> {
+    return this.boardRepository.createBoard(createBoardDto, user);
   }
 
   // ID로 게시물 가져오기
-  async getBoardById(id: number): Promise<Board> {
-    const found = await this.boardRepository.findOne(id);
+  async getBoardById(id: number, user: User): Promise<Board> {
+    const found = await this.boardRepository.findOne({
+      where: { id, userId: user.id },
+    });
 
     if (!found) {
       throw new NotFoundException(`Can't find Board with id ${id}`);
@@ -36,8 +42,8 @@ export class BoardsService {
   }
 
   // ID로 게시물 삭제하기
-  async deleteBoard(id: number): Promise<void> {
-    const result = await this.boardRepository.delete(id);
+  async deleteBoard(id: number, user: User): Promise<void> {
+    const result = await this.boardRepository.delete({ id, userId: user.id });
 
     if (result.affected === 0) {
       throw new NotFoundException(`Can't find Board with id ${id}`);
@@ -45,8 +51,12 @@ export class BoardsService {
   }
 
   // ID로 게시물 상태 수정하기
-  async updateBoardStatus(id: number, status: BoardStatus): Promise<Board> {
-    const board = await this.getBoardById(id);
+  async updateBoardStatus(
+    id: number,
+    user: User,
+    status: BoardStatus,
+  ): Promise<Board> {
+    const board = await this.getBoardById(id, user);
 
     board.status = status;
     await this.boardRepository.save(board);
